remove hail from current commit. seperate plink dataset into seperate file.
